requirements.txt
# Core ML and AI
tensorflow==2.13.0
torch==2.0.1
transformers==4.31.0
openai==0.27.8
langchain==0.0.262
sentence-transformers==2.2.2

# Data Processing
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
opencv-python==4.8.0.74

# Deep Learning & Vision
torchvision==0.15.2
pillow==10.0.0

# NLP
spacy==3.6.0
nltk==3.8.1

# API and Web
fastapi==0.100.0
uvicorn==0.23.1
python-multipart==0.0.6

# Database
pymongo==4.4.1
redis==4.6.0

# Monitoring and MLOps
mlflow==2.6.0
wandb==0.15.5
prometheus-client==0.17.1

# Testing and Quality
pytest==7.4.0
black==23.7.0
flake8==6.0.0

# Documentation
sphinx==7.0.1
mkdocs==1.5.2

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
tqdm==4.65.0 